+++ test/unit/clear.js
@@ -6,21 +6,7 @@ var manager = require('../manager');
 
 describe('clear(cb)', function() {
 
-	var sessionStore;
-
-	before(function(done) {
-
-		manager.setUp(function(error, store) {
-
-			if (error) {
-				return done(error);
-			}
-
-			sessionStore = store;
-			done();
-		});
-	});
-
+	before(manager.setUp);
 	after(manager.tearDown);
 
 	describe('when sessions exist', function() {
@@ -29,12 +15,9 @@ describe('clear(cb)', function() {
 
 		it('should delete all existing sessions', function(done) {
 
-			sessionStore.clear(function(error) {
-
+			manager.sessionStore.clear(function(error) {
 				expect(error).to.be.undefined;
-
-				sessionStore.length(function(error, count) {
-
+				manager.sessionStore.length(function(error, count) {
 					expect(error).to.equal(null);
 					expect(count).to.equal(0);
 					done();