+++ test/unit/close.js
@@ -14,21 +14,14 @@ describe('close(cb)', function() {
 		describe('default options', function() {
 
 			var sessionStore;
-
 			beforeEach(function(done) {
 				sessionStore = new MySQLStore(manager.config, done);
 			});
 
 			it('should close the store and end the database connection', function(done) {
-
 				sessionStore.close(function(error) {
-
-					if (error) {
-						return done(error);
-					}
-
+					if (error) return done(error);
 					expect(sessionStore._expirationInterval).to.equal(null);
-
 					sessionStore.connection.query('SHOW TABLES', function(error, result) {
 						expect(error).to.not.equal(null);
 						expect(error.code).to.equal('POOL_CLOSED');
@@ -41,13 +34,10 @@ describe('close(cb)', function() {
 		describe('option.endConnectionOnClose set to FALSE', function() {
 
 			var sessionStore;
-
 			beforeEach(function(done) {
-
 				var options = _.extend({}, manager.config, {
 					endConnectionOnClose: false
 				});
-
 				sessionStore = new MySQLStore(options, done);
 			});
 
@@ -56,15 +46,9 @@ describe('close(cb)', function() {
 			});
 
 			it('should close the store but not end the database connection', function(done) {
-
 				sessionStore.close(function(error) {
-
-					if (error) {
-						return done(error);
-					}
-
+					if (error) return done(error);
 					expect(sessionStore._expirationInterval).to.equal(null);
-
 					sessionStore.connection.query('SHOW TABLES', function(error, result) {
 						expect(error).to.equal(null);
 						done();
@@ -79,7 +63,6 @@ describe('close(cb)', function() {
 		describe('default options', function() {
 
 			var sessionStore;
-
 			beforeEach(function(done) {
 				var connection = mysql.createPool(manager.config);
 				sessionStore = new MySQLStore({}/* options */, connection, done);
@@ -90,15 +73,9 @@ describe('close(cb)', function() {
 			});
 
 			it('should close the store but not end the database connection', function(done) {
-
 				sessionStore.close(function(error) {
-
-					if (error) {
-						return done(error);
-					}
-
+					if (error) return done(error);
 					expect(sessionStore._expirationInterval).to.equal(null);
-
 					sessionStore.connection.query('SHOW TABLES', function(error, result) {
 						expect(error).to.equal(null);
 						done();
@@ -110,26 +87,17 @@ describe('close(cb)', function() {
 		describe('option.endConnectionOnClose set to TRUE', function() {
 
 			var sessionStore;
-
 			beforeEach(function(done) {
-
 				var options = _.extend({}, manager.config, {
 					endConnectionOnClose: true
 				});
-
 				sessionStore = new MySQLStore(options, done);
 			});
 
 			it('should close the store and end the database connection', function(done) {
-
 				sessionStore.close(function(error) {
-
-					if (error) {
-						return done(error);
-					}
-
+					if (error) return done(error);
 					expect(sessionStore._expirationInterval).to.equal(null);
-
 					sessionStore.connection.query('SHOW TABLES', function(error, result) {
 						expect(error).to.not.equal(null);
 						expect(error.code).to.equal('POOL_CLOSED');