+++ test/unit/touch.js
@@ -8,15 +8,7 @@ var fixtures = manager.fixtures.sessions;
 
 describe('touch(session_id, data, cb)', function() {
 
-	var sessionStore;
-	before(function(done) {
-		manager.setUp(function(error, store) {
-			if (error) return done(error);
-			sessionStore = store;
-			done();
-		});
-	});
-
+	before(manager.setUp);
 	after(manager.tearDown);
 
 	describe('when the session does not exist', function() {
@@ -26,27 +18,17 @@ describe('touch(session_id, data, cb)', function() {
 		it('should not create new session', function(done) {
 
 			async.each(fixtures, function(fixture, nextFixture) {
-
 				var session_id = fixture.session_id;
 				var data = fixture.data;
-
-				sessionStore.touch(session_id, data, function(error) {
-
+				manager.sessionStore.touch(session_id, data, function(error) {
 					expect(error).to.be.undefined;
-
-					sessionStore.get(session_id, function(error, session) {
-
-						if (error) {
-							return nextFixture(error);
-						}
-
+					manager.sessionStore.get(session_id, function(error, session) {
+						if (error) return nextFixture(error);
 						expect(error).to.equal(null);
 						expect(session).to.equal(null);
-
 						nextFixture();
 					});
 				});
-
 			}, done);
 		});
 	});
@@ -56,13 +38,10 @@ describe('touch(session_id, data, cb)', function() {
 		var oldExpiresValue = Math.round((Date.now() / 1000)) - 10;
 
 		before(function(done) {
-
 			manager.populateSessions(function() {
-
 				var sql = 'UPDATE `sessions` SET `expires` = ?';
 				var params = [oldExpiresValue];
-
-				sessionStore.connection.query(sql, params, done);
+				manager.sessionStore.connection.query(sql, params, done);
 			});
 		});
 
@@ -72,7 +51,7 @@ describe('touch(session_id, data, cb)', function() {
 
 				var session_id = fixture.session_id;
 
-				sessionStore.touch(session_id, fixture.data, function(error) {
+				manager.sessionStore.touch(session_id, fixture.data, function(error) {
 
 					expect(error).to.be.undefined;
 
@@ -82,14 +61,11 @@ describe('touch(session_id, data, cb)', function() {
 						fixture.session_id
 					];
 
-					sessionStore.connection.query(sql, params, function(error, data) {
-
+					manager.sessionStore.connection.query(sql, params, function(error, data) {
 						var touchedSession = data[0];
-
 						expect(touchedSession.session_id).to.equal(fixture.session_id);
 						expect(touchedSession.data).to.equal(JSON.stringify(fixture.data));
 						expect(touchedSession.expires).to.above(oldExpiresValue);
-
 						nextFixture();
 					});
 				});